tests/wpt/web-platform-tests/app-history/navigate/goTo-detach-between-navigate-and-navigatesuccess.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="return-value/resources/helpers.js"></script>

<iframe id="i" src="/common/blank.html"></iframe>

<script>
async_test(t => {
  window.onload = t.step_func(() => {
    const iWindow = i.contentWindow;
    const iDOMException = iWindow.DOMException;
    const iAppHistoryEntry = iWindow.AppHistoryEntry;

    i.contentWindow.appHistory.navigate("#1").finished.then(t.step_func(() => {
      assert_equals(i.contentWindow.appHistory.entries().length, 2);
      let key = i.contentWindow.appHistory.entries()[0].key;

      let onnavigateerror_called = false;
      let result;

      i.contentWindow.appHistory.onnavigate = t.step_func(e => {
        e.transitionWhile(new Promise(resolve => () => t.step_timeout(resolve, 2)));
        t.step_timeout(() => i.remove(), 1);
      });

      i.contentWindow.appHistory.onnavigatesuccess = t.unreached_func("navigatesuccess must not fire");

      i.contentWindow.appHistory.onnavigateerror = t.step_func(e => {
        assert_false(onnavigateerror_called);
        onnavigateerror_called = true;
        assert_equals(e.filename, location.href);
        assert_greater_than(e.lineno, 0);
        assert_greater_than(e.colno, 0);

        assertCommittedFulfillsFinishedRejectsExactly(t, result, iWindow.appHistory.current, e.error, iWindow, iDOMException, iAppHistoryEntry).then(
          () => t.done(),
          t.step_func(err => { throw err; })
        );
      });

      result = i.contentWindow.appHistory.goTo(key);
    }));
  });
}, "Detach a window between when a goTo() fires navigate and navigatesuccess");
</script>
